Updated some attributes, and added the main contract in which the game should be hold.

+++ src/card-game.sol
@@ -3,7 +3,7 @@ pragma solidity >=0.4.16 <0.7.0;
 interface CardGame {
     
 
-    function changeCardSuit(uint16 _suit) external returns (uint8);
+    function changeCardSuit(uint16 _suit) external returns (uint16);
 
     function revealCardValue(uint16 _index) external view returns(uint16);
 
@@ -12,4 +12,5 @@ interface CardGame {
     function revealFullCard(uint16 _index) external view returns(uint16,uint16);
 
     function commitCard(uint16 _card_committed, uint16 _quantity) external;
+
 }
